Worked on tests

+++ pyfvde/pyfvde_datetime.h
@@ -41,7 +41,7 @@ PyObject *pyfvde_datetime_new_from_floatingtime(
            uint64_t floatingtime );
 
 PyObject *pyfvde_datetime_new_from_posix_time(
-           uint32_t posix_time );
+           int64_t posix_time );
 
 PyObject *pyfvde_datetime_new_from_posix_time_in_micro_seconds(
            int64_t posix_time );
